Rename new MetricsPage module package
